feat: now navbar email are provided by jwt token
