+++ common/types/Maybe.ts
@@ -1,0 +1,1 @@
+﻿export type Maybe<T> = T | null | undefined;
