+++ frontend/app/home/components/RingsList/RingsList.tsx
